requirements.txt
aiogram-3.20.0
python-dotenv-1.1.0
aiosqlite-0.21.0